Rename files in tasks module and move tasks related models and routes to the tasks directory. This closes #15.
