+++ backend/app/tasks/routers.py
@@ -1,21 +1,22 @@
 from fastapi import APIRouter, HTTPException, status, Depends
 from pydantic import UUID4
-from app.tasks import schema as tasks_schema
-from app import models
 from app.database import get_db
 from sqlalchemy.orm import Session
 from sqlalchemy.exc import IntegrityError
 from typing import List
 
+from app.tasks.schemas import TaskRead, TaskCreate
+from app.tasks.models import Task
+
 
 router = APIRouter()
 
 
-@router.post("/", response_model=tasks_schema.TaskRead)
-def create_task(task: tasks_schema.TaskCreate, db: Session = Depends(get_db)):
+@router.post("/", response_model=TaskRead)
+def create_task(task: TaskCreate, db: Session = Depends(get_db)):
     """API endpoint for adding a task"""
 
-    new_task = models.Task(**task.dict())
+    new_task = Task(**task.dict())
 
     try:
         db.add(new_task)
@@ -31,20 +32,20 @@ def create_task(task: tasks_schema.TaskCreate, db: Session = Depends(get_db)):
     return new_task
 
 
-@router.get("/", response_model=List[tasks_schema.TaskRead])
+@router.get("/", response_model=List[TaskRead])
 def get_all_tasks(db: Session = Depends(get_db)):
     """API endpoint for getting all tasks"""
 
-    tasks = db.query(models.Task).all()
+    tasks = db.query(Task).all()
 
     return tasks
 
 
-@router.get("/{uuid}", response_model=tasks_schema.TaskRead)
+@router.get("/{uuid}", response_model=TaskRead)
 def get_task(uuid: UUID4, db: Session = Depends(get_db)):
     """API endpoint to get an individual task by its uuid"""
 
-    task = db.query(models.Task).filter(models.Task.id == uuid).first()
+    task = db.query(Task).filter(Task.id == uuid).first()
 
     if task is None:
         raise HTTPException(
@@ -59,7 +60,7 @@ def get_task(uuid: UUID4, db: Session = Depends(get_db)):
 def delete_task(uuid: UUID4, db: Session = Depends(get_db)):
     """API endpoint to delete a task"""
 
-    task = db.query(models.Task).filter(models.Task.id == uuid)
+    task = db.query(Task).filter(Task.id == uuid)
 
     if task.first() is None:
         raise HTTPException(
@@ -72,15 +73,15 @@ def delete_task(uuid: UUID4, db: Session = Depends(get_db)):
     db.commit()
 
 
-@router.put("/{uuid}", response_model=tasks_schema.TaskRead)
+@router.put("/{uuid}", response_model=TaskRead)
 def update_task(
     uuid: UUID4,
-    updated_task: tasks_schema.TaskCreate,
+    updated_task: TaskCreate,
     db: Session = Depends(get_db),
 ):
     """API endpoint to update a task"""
 
-    task_query = db.query(models.Task).filter(models.Task.id == uuid)
+    task_query = db.query(Task).filter(Task.id == uuid)
 
     task = task_query.first()
